Remove `Population` class, Create population class variable in `Grid`
Removed the `Population` class. The population is now a class variable
of `Agent`s that belong to the `Grid` class.

I also created a new function, `createPopulation(n: number)` that
accepts the size of the population.

The size of the population should be linked to the value of a slider
that the user control. This still has to be implemented.